+++ config/Includes.json
@@ -2,5 +2,5 @@
 	["Unirest.php", true],
 	["API_Wikipedia.php", true],
 	["API_Google.php", true],
-	["API_Youtube.php", true]
+	["Youtube/Youtube.php", true]
 ]
